event/Loop: don't wait for io_uring if epoll is still ready
This fixes a starvation bug if epoll is getting hammered with
completions.  This fills the io_uring completion queue until it
overflows, but these overflows are never picked up because we never
invoke io_uring_enter().

We avoid this by calling io_uring_enter() in any case, but with a zero
timeout.